azure-entra-agent/pyproject.toml
[project]
name = "azure-entra-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.1",
    "msgraph-sdk>=1.49.0",
    "python-dotenv>=1.2.1",
]
